README.md
This is a little Brain teaser console game. “Where do words go?” was created in 2014 by Dr. Pascale Michelona to exercise your cognitive skills. Aims to stimulate the connections or associations between words in your temporal lobe. You will see pairs of words, and your goal is to find a third word that is connected or associated with both two words.

https://sharpbrains.com/blog/2014/06/20/brain-teaser-to-exercise-your-cognitive-skills-where-do-words-go/